Move stream to raw as most users shouldn't look at it
